+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - [Library] `password input` component (Orange-OpenSource/ouds-ios#997)
+- [Library] `bullet list` component (Orange-OpenSource/ouds-ios#513)
 
 ### Changed